Move get_spectral_geometry to epygram.spectra
